refactor(converter): split job message catalog and mapper
